refactor: rename `escape` to `runtimeEscape`
