+++ lib/loader.js
@@ -113,10 +113,10 @@ module.exports = function(content, map) {
 
       if (query.url !== false && urlItems.length > 0) {
         urlEscapeHelper =
-          "var escape = require(" +
+          "var runtimeEscape = require(" +
           loaderUtils.stringifyRequest(
             loaderContext,
-            require.resolve("./escape.js")
+            require.resolve("./runtimeEscape.js")
           ) +
           ");\n";
 
@@ -142,7 +142,7 @@ module.exports = function(content, map) {
               urlRequest = url.substr(0, idx);
 
               return (
-                '" + escape(require(' +
+                '" + runtimeEscape(require(' +
                 loaderUtils.stringifyRequest(loaderContext, urlRequest) +
                 ')) + "' +
                 url.substr(idx)
@@ -152,7 +152,7 @@ module.exports = function(content, map) {
             urlRequest = url;
 
             return (
-              '" + escape(require(' +
+              '" + runtimeEscape(require(' +
               loaderUtils.stringifyRequest(loaderContext, urlRequest) +
               ')) + "'
             );
