+++ lib/runtimeEscape.js
