+++ test/helpers.js
@@ -9,10 +9,11 @@ function getEvaluated(output, modules) {
 		var fn = vm.runInThisContext("(function(module, exports, require) {" + output + "})", "testcase.js");
 		var m = { exports: {}, id: 1 };
 		fn(m, m.exports, function(module) {
-			if(module.indexOf("runtime") >= 0)
+			if(module.indexOf("runtime.js") >= 0)
 				return require("../lib/runtime");
-			if(module.indexOf("escape") >= 0)
-				return require("../lib/escape");
+			if(module.indexOf("runtimeEscape.js") >= 0) {
+        return require("../lib/runtimeEscape");
+      }
 			if(module.indexOf("-!/path/css-loader!") === 0)
 				module = module.substr(19);
 			if(modules && module in modules)
